# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "CoinDirection Integration APIs",
    "version": "1.0.0",
    "title": "Swagger CoinDirection Integration APIs"
  },
  "host": "localhost:3000",
  "tags": [
    {
      "name": "countries",
      "description": "Swagger CoinDirection Integration APIs"
    }
  ],
  "paths": {
    "/countries": {
      "get": {
        "tags": [
          "countries  "
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name of the country",
            "type": "string"
          },
          {
            "name": "currency",
            "in": "query",
            "type": "string",
            "description": "Currency used in the country"
          },
          {
            "name": "offset",
            "in": "query",
            "type": "number",
            "description": "Start offset for pagination"
          },
          {
            "name": "maxresult",
            "in": "query",
            "type": "number",
            "description": "Maximum result expected in the response : support for pagination"
          },
          {
            "name": "sortOrderOnCurrency",
            "in": "query",
            "description": "Sort countries based on currencies",
            "type": "array",
            "items" :{
              "type" : "string",
              "enum" : [
                "ASC", "DSC"
              ]
            }
          },
          {
            "name": "sortOrderOnName",
            "in": "query",
            "description": "Sort countries based on county names",
            "type": "array",
            "items" :{
              "type" : "string",
              "enum" : [
                "ASC", "DSC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Country"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/currencies": {
      "get": {
        "tags": [
          "currencies"
        ],
        "produces": [
          "application/json"
        ],

        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Currency"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    }
  },
  "definitions": {
    "Country": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "documents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/document"
          }
        },
        "maxWithdrawl": {
          "type": "number"
        }
      }
    },
    "document": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "code": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "required": {
          "type": "string"
        }
      }
    },
    "Currency": {
      "type": "object",
      "properties": {
        "currency": {
          "type": "string"
        },

        "countries": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }

      }
    }
  }
}